requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
docopt==0.6.2
fonttools==4.38.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.0
nptyping==2.3.1
numpy==1.23.4
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.1
patsy==0.5.3
Pillow==9.2.0
pipreqs==0.4.11
pynrrd==1.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
statsmodels==0.13.2
threadpoolctl==3.1.0
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
typing==3.7.4.3
typing_extensions==4.4.0
urllib3==1.26.12
yarg==0.1.9
